Moving handler function to Services
